+++ src/components/parts/Body/styles.ts
