+++ src/components/patterns/Footer/index.tsx
@@ -4,8 +4,8 @@ import { Path } from '@/app/constants/paths';
 import { URLS } from '@/app/constants/urls';
 import { useResponsive } from '@/app/hooks/useResponsive';
 import { SecondaryButton } from '@/components/parts/Button/SecondaryButton';
-import { InternalLink } from '@/components/parts/Text/InternalLink';
-import { Label } from '@/components/parts/Text/Label';
+import { InternalLink } from '@/components/parts/InternalLink';
+import { Label } from '@/components/parts/Label';
 import { QRBalloon } from '@/components/patterns/Balloon/QRBalloon';
 
 import { NAV_LABELS, NAV_LIST } from './navData';
