+++ src/components/patterns/Header/MobileLayout/index.tsx
@@ -3,8 +3,8 @@ import React, { useState } from 'react';
 import { Path } from '@/app/constants/paths';
 import { Container } from '@/components/layouts/Container';
 import { Icon } from '@/components/parts/Image/Icon';
-import { InternalLink } from '@/components/parts/Text/InternalLink';
-import { Label } from '@/components/parts/Text/Label';
+import { InternalLink } from '@/components/parts/InternalLink';
+import { Label } from '@/components/parts/Label';
 
 import { styles, getStyles } from './styles';
