+++ src/components/patterns/Modal/PreviewModal/index.tsx
@@ -3,7 +3,7 @@ import React, { SetStateAction, useCallback, useMemo } from 'react';
 import { ModalWrapper } from '@/components/libs/Modal/ModalWrapper';
 import { ActionButton } from '@/components/parts/Button/ActionButton';
 import { Icon } from '@/components/parts/Image/Icon';
-import { Label } from '@/components/parts/Text/Label';
+import { Label } from '@/components/parts/Label';
 
 import { styles, getButtonWrapperStyle, imageStyle } from './styles';
